# openapi/spec.yml
openapi: 3.0.0
info:
  contact:
    name: Canva Developer Community
    url: https://community.canva.dev/
  description: API for building integrations with Canva via a REST api
  license:
    name: ©2023 All Rights Reserved
  termsOfService: https://www.canva.com/trust/legal/
  title: Canva Connect API
  version: latest
servers:
- description: Canva Connect API
  url: https://api.canva.com/rest
...
paths:
...
  /v1/exports:
    post:
      description: |-
        Starts a new job to export a file from Canva. Once the exported file is generated, you can download
        it using the link(s) provided.

        The request requires the design ID and the exported file format.

        Supported formats: pdf(`.pdf`), jpg(`.jpg`), png(`.png`), gif(`.gif`),
        pptx(`.pptx`), and mp4(`.mp4`).

        You can check the status and get the results of export jobs created with this API using the [Get design export job API](https://www.canva.dev/docs/connect/api-reference/exports/get-design-export-job/).

        NOTE: If you set `export_quality` to `pro`, the export might fail if the design contains [premium elements](https://www.canva.com/help/premium-elements/), and the calling user either hasn't purchased the elements or isn't on a Canva plan (such as Canva Pro) that has premium features.
      operationId: createDesignExportJob
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDesignExportJobRequest'
      responses:
        "200":
          content:
            application/json:
              examples:
                in_progress:
                  $ref: '#/components/examples/InProgressExportJobExample'
                success:
                  $ref: '#/components/examples/SuccessExportJobExample'
                failed:
                  $ref: '#/components/examples/FailedExportJobExample'
              schema:
                $ref: '#/components/schemas/CreateDesignExportJobResponse'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error Response
      security:
      - oauthAuthCode:
        - design:content:read
      tags:
      - export
      x-rate-limit-per-client-user: 20
  /v1/exports/{exportId}:
    get:
      description: |-
        Gets the result of a design export job that was created using the [Create design export job API](https://www.canva.dev/docs/connect/api-reference/exports/create-design-export-job/).

        If the job is complete, the response includes an array
        of download links for each page of the design.
      operationId: getDesignExportJob
      parameters:
      - description: The export job ID.
        explode: false
        in: path
        name: exportId
        required: true
        schema:
          pattern: "^[a-zA-Z0-9_-]{1,50}$"
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              examples:
                in_progress:
                  $ref: '#/components/examples/InProgressExportJobExample'
                success:
                  $ref: '#/components/examples/SuccessExportJobExample'
                failed:
                  $ref: '#/components/examples/FailedExportJobExample'
              schema:
                $ref: '#/components/schemas/GetDesignExportJobResponse'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error Response
      security:
      - oauthAuthCode:
        - design:content:read
      tags:
      - export
      x-rate-limit-per-client-user: 120
...
components:
  examples:
...
    InProgressExportJobExample:
      summary: In progress job
      value:
        job:
          id: e08861ae-3b29-45db-8dc1-1fe0bf7f1cc8
          status: in_progress
    SuccessExportJobExample:
      summary: Successfully completed job
      value:
        job:
          id: e08861ae-3b29-45db-8dc1-1fe0bf7f1cc8
          status: success
          urls:
          - https://export-download.canva.com/...
    FailedExportJobExample:
      summary: Failed job
      value:
        job:
          id: e08861ae-3b29-45db-8dc1-1fe0bf7f1cc8
          status: failed
          error:
            code: license_required
            message: User doesn't have the required license to export in PRO quality.
...
  schemas:
...
    Error:
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          description: A human-readable description of what went wrong.
          type: string
      required:
      - code
      - message
      type: object
...
    ErrorCode:
      description: |
        A short string indicating what failed. This field can be used to handle errors programmatically.
      enum:
      - internal_error
      - invalid_field
      - invalid_header_value
      - permission_denied
      - too_many_requests
      - not_found
      - bad_request_body
      - bad_http_method
      - bad_request_params
      - bad_query_params
      - endpoint_not_found
      - unsupported_version
      - invalid_access_token
      - revoked_access_token
      - missing_field
      - missing_scope
      - invalid_grant
      - invalid_request
      - invalid_client
      - unauthorized_client
      - unsupported_grant_type
      - invalid_scope
      - invalid_basic_header
      - invalid_file_format
      - quota_exceeded
      - unsupported_content_type
      - request_too_large
      - folder_not_found
      - asset_not_found
      - max_limit_reached
      - permission_not_found
      - permission_exists
      - unauthorized_user
      - user_not_found
      - group_not_found
      - app_not_found
      - content_not_found
      - doctype_not_found
      - design_not_found
      - design_type_not_found
      - team_not_found
      - comment_not_found
      - too_many_comments
      - too_many_replies
      - message_too_long
      - design_not_fillable
      - autofill_data_invalid
      - feature_not_available
      - license_required
      type: string
    CreateDesignExportJobRequest:
      description: |-
        Body parameters for starting an export job for a design.
        It must include a design ID, and one of the supported export formats.
      example:
        design_id: DAVZr1z5464
        format:
          type: pdf
          size: a4
          pages:
          - 2
          - 3
          - 4
      properties:
        design_id:
          description: The design ID.
          type: string
        format:
          $ref: '#/components/schemas/ExportFormat'
      required:
      - design_id
      - format
      type: object
    ExportFormat:
      description: Details about the desired export format.
      discriminator:
        mapping:
          pdf: '#/components/schemas/PdfExportFormat'
          jpg: '#/components/schemas/JpgExportFormat'
          png: '#/components/schemas/PngExportFormat'
          pptx: '#/components/schemas/PptxExportFormat'
          gif: '#/components/schemas/GifExportFormat'
          mp4: '#/components/schemas/Mp4ExportFormat'
        propertyName: type
      oneOf:
      - $ref: '#/components/schemas/PdfExportFormat'
      - $ref: '#/components/schemas/JpgExportFormat'
      - $ref: '#/components/schemas/PngExportFormat'
      - $ref: '#/components/schemas/PptxExportFormat'
      - $ref: '#/components/schemas/GifExportFormat'
      - $ref: '#/components/schemas/Mp4ExportFormat'
      type: object
    PdfExportFormat:
      description: Export the design as a PDF. Providing a paper size is optional.
        The default paper size is A4.
      properties:
        type:
          enum:
          - pdf
          type: string
        export_quality:
          $ref: '#/components/schemas/ExportQuality'
        size:
          $ref: '#/components/schemas/ExportPageSize'
        pages:
          description: |-
            To specify which pages to export in a multi-page design, provide the page numbers as
            an array. The first page in a design is page `1`.
            If `pages` isn't specified, all the pages are exported.
          example:
          - 2
          - 3
          - 4
          items:
            minimum: 1
            type: integer
          type: array
      required:
      - type
      type: object
    GifExportFormat:
      description: |-
        Export the design as a GIF. Height or width (or both) may be specified, otherwise the file
        will be exported at it's default size. Large designs will be scaled down, and aspect ratio
        will always be maintained.
      properties:
        type:
          enum:
          - gif
          type: string
        export_quality:
          $ref: '#/components/schemas/ExportQuality'
        height:
          description: |-
            Specify the height in pixels of the exported image. If only one of height or width is
            specified, then the image will be scaled to match that dimension, respecting the design's
            aspect ratio. If no width or height is specified, the image will be exported using the
            dimensions of the design.
          example: 400
          format: int32
          maximum: 25000
          minimum: 40
          type: integer
        width:
          description: |-
            Specify the width in pixels of the exported image. If only one of height or width is
            specified, then the image will be scaled to match that dimension, respecting the design's
            aspect ratio. If no width or height is specified, the image will be exported using the
            dimensions of the design.
          example: 400
          format: int32
          maximum: 25000
          minimum: 40
          type: integer
        pages:
          description: |-
            To specify which pages to export in a multi-page design, provide the page numbers as
            an array. The first page in a design is page `1`.
            If `pages` isn't specified, all the pages are exported.
          example:
          - 2
          - 3
          - 4
          items:
            minimum: 1
            type: integer
          type: array
      required:
      - type
      type: object
    JpgExportFormat:
      description: |-
        Export the design as a JPEG. Compression quality must be provided. Height or width (or both)
        may be specified, otherwise the file will be exported at it's default size.

        If the user is on the Canva Free plan, the export height and width for a fixed-dimension design can't be upscaled by more than a factor of `1.125`.
      properties:
        type:
          enum:
          - jpg
          type: string
        export_quality:
          $ref: '#/components/schemas/ExportQuality'
        quality:
          description: "For the `jpg` type, the `quality` of the exported JPEG determines\
            \ how compressed the exported file should be. A _low_ `quality` value\
            \ (minimum `1`) will create a file with a smaller file size, but the resulting\
            \ file will have pixelated artifacts when compared to a file created with\
            \ a _high_ `quality` value (maximum `100`)."
          example: 80
          maximum: 100
          minimum: 1
          type: integer
        height:
          description: |-
            Specify the height in pixels of the exported image. If only one of height or width is
            specified, then the image will be scaled to match that dimension, respecting the design's
            aspect ratio. If no width or height is specified, the image will be exported using the
            dimensions of the design.
          example: 400
          format: int32
          maximum: 25000
          minimum: 40
          type: integer
        width:
          description: |-
            Specify the width in pixels of the exported image. If only one of height or width is
            specified, then the image will be scaled to match that dimension, respecting the design's
            aspect ratio. If no width or height is specified, the image will be exported using the
            dimensions of the design.
          example: 400
          format: int32
          maximum: 25000
          minimum: 40
          type: integer
        pages:
          description: |-
            To specify which pages to export in a multi-page design, provide the page numbers as
            an array. The first page in a design is page `1`.
            If `pages` isn't specified, all the pages are exported.
          example:
          - 2
          - 3
          - 4
          items:
            minimum: 1
            type: integer
          type: array
      required:
      - quality
      - type
      type: object
    PngExportFormat:
      description: |-
        Export the design as a PNG. Height or width (or both) may be specified, otherwise
        the file will be exported at it's default size. You may also specify whether to export the
        file losslessly, and whether to export a multi-page design as a single image.

        If the user is on the Canva Free plan, the export height and width for a fixed-dimension design can't be upscaled by more than a factor of `1.125`.
      properties:
        type:
          enum:
          - png
          type: string
        export_quality:
          $ref: '#/components/schemas/ExportQuality'
        height:
          description: |-
            Specify the height in pixels of the exported image. If only one of height or width is
            specified, then the image will be scaled to match that dimension, respecting the design's
            aspect ratio. If no width or height is specified, the image will be exported using the
            dimensions of the design.
          example: 400
          format: int32
          maximum: 25000
          minimum: 40
          type: integer
        width:
          description: |-
            Specify the width in pixels of the exported image. If only one of height or width is
            specified, then the image will be scaled to match that dimension, respecting the design's
            aspect ratio. If no width or height is specified, the image will be exported using the
            dimensions of the design.
          example: 400
          format: int32
          maximum: 25000
          minimum: 40
          type: integer
        lossless:
          default: false
          description: |-
            When `true`, the PNG is compressed with a lossless compression algorithm (`false` by
            default).
          type: boolean
        as_single_image:
          default: false
          description: |-
            When `true`, multi-page designs are merged into a single image.
            When `false` (default), each page is exported as a separate image.
          type: boolean
        pages:
          description: |-
            To specify which pages to export in a multi-page design, provide the page numbers as
            an array. The first page in a design is page `1`.
            If `pages` isn't specified, all the pages are exported.
          example:
          - 2
          - 3
          - 4
          items:
            minimum: 1
            type: integer
          type: array
      required:
      - type
      type: object
    PptxExportFormat:
      description: Export the design as a PPTX.
      properties:
        type:
          enum:
          - pptx
          type: string
        pages:
          description: |-
            To specify which pages to export in a multi-page design, provide the page numbers as
            an array. The first page in a design is page `1`.
            If `pages` isn't specified, all the pages are exported.
          example:
          - 2
          - 3
          - 4
          items:
            minimum: 1
            type: integer
          type: array
      required:
      - type
      type: object
    Mp4ExportFormat:
      description: Export the design as an MP4. You must specify the quality of the
        exported video.
      properties:
        type:
          enum:
          - mp4
          type: string
        export_quality:
          $ref: '#/components/schemas/ExportQuality'
        quality:
          $ref: '#/components/schemas/Mp4ExportQuality'
        pages:
          description: |-
            To specify which pages to export in a multi-page design, provide the page numbers as
            an array. The first page in a design is page `1`.
            If `pages` isn't specified, all the pages are exported.
          example:
          - 2
          - 3
          - 4
          items:
            minimum: 1
            type: integer
          type: array
      required:
      - quality
      - type
      type: object
    CreateDesignExportJobResponse:
      properties:
        job:
          $ref: '#/components/schemas/ExportJob'
      required:
      - job
      type: object
    GetDesignExportJobResponse:
      properties:
        job:
          $ref: '#/components/schemas/ExportJob'
      required:
      - job
      type: object
    ExportJob:
      description: The status of the export job.
      properties:
        id:
          description: The export job ID.
          example: e08861ae-3b29-45db-8dc1-1fe0bf7f1cc8
          type: string
        status:
          $ref: '#/components/schemas/DesignExportStatus'
        urls:
          description: |-
            When the export job is completed, also returns a list of urls for the exported
            resources. The list is sorted by page order.
          example:
          - https://export-download.canva-dev.com/...
          items:
            type: string
          type: array
        error:
          $ref: '#/components/schemas/ExportError'
      required:
      - id
      - status
      type: object
    ExportPageSize:
      description: The paper size of the export PDF file. The `size` attribute is
        only supported for Documents (Canva Docs).
      enum:
      - a4
      - a3
      - letter
      - legal
      example: a4
      type: string
    DesignExportStatus:
      description: |-
        The export status of the job. A newly created job will be `in_progress` and will eventually
        become `success` or `failed`.
      enum:
      - failed
      - in_progress
      - success
      type: string
...
    Mp4ExportQuality:
      description: |-
        The orientation and resolution of the exported video. Orientation is either `horizontal` or
        `vertical`, and resolution is one of `480p`, `720p`, `1080p` or `4k`.
      enum:
      - horizontal_480p
      - horizontal_720p
      - horizontal_1080p
      - horizontal_4k
      - vertical_480p
      - vertical_720p
      - vertical_1080p
      - vertical_4k
      type: string
    ExportError:
      description: "If the export fails, this object provides details about the error."
      properties:
        code:
          $ref: '#/components/schemas/ExportErrorCode'
        message:
          description: A human-readable description of what went wrong.
          type: string
      type: object
    ExportErrorCode:
      description: |-
        If the export failed, this specifies the reason why it failed. This can be one of the following:

        - `license_required`: The design contains [premium elements](https://www.canva.com/help/premium-elements/) that haven't been purchased. You can either buy the elements or upgrade to a Canva plan (such as Canva Pro) that has premium features, then try again. Alternatively, you can set `export_quality` to `regular` to export your document in regular quality.
        - `internal_failure`: The service encountered an error when exporting your design.
      enum:
      - license_required
      - internal_failure
      type: string
    ExportQuality:
      default: regular
      description: |-
        Specifies the export quality of the design. This can be one of the following:
        - `regular`: Regular quality export. This is the default.
        - `pro`: Premium quality export.

          NOTE: A `pro` export might fail if the design contains [premium elements](https://www.canva.com/help/premium-elements/) and the calling user either hasn't purchased the elements or isn't on a Canva plan (such as Canva Pro) that has premium features.
      enum:
      - pro
      - regular
      type: string
...
  securitySchemes:
    basicAuth:
      scheme: basic
      type: http
    oauthAuthCode:
      flows:
        authorizationCode:
          authorizationUrl: https://www.canva.com/api/oauth/authorize
          scopes:
            design:content:read: View the contents of the user's designs.
            design:meta:read: View the metadata of the user's designs.
            design:content:write: Create designs on the user's behalf.
            folder:read: |-
              View the metadata and contents of the user's folders, including their **Projects**
              folder.
            folder:write: |-
              Add, move, or remove the user's folders. It also lets you edit folder metadata,
              such as the folder's name.
            asset:read: "View the metadata for the user's assets, such as uploaded\
              \ images."
            asset:write: "Upload, update, or delete assets on the user's behalf."
            comment:read: "View the comments on the user's designs, and the associated\
              \ metadata."
            comment:write: Create comments and replies on the user's designs.
            collaboration:event: Receive webhook notifications about events relevant
              to the user.
            brandtemplate:meta:read: View the metadata of the brand templates associated
              with the user's brand.
            brandtemplate:content:read: Read the content of the brand templates associated
              with the user's brand.
            profile:read: Read a user's profile and account information.
          tokenUrl: https://api.canva.com/rest/v1/oauth/token
      type: oauth2